tungsten_ci_utils/test_analyzer/requirements.txt
lxml==4.2.3
peewee==3.5.2
PyYAML
